\name{convmtx}
\alias{convmtx}
\title{
Convolution matrix
}
\description{
Returns the convolution matrix, A, associated to the filter v such that the product of A and an n-element vector, x, is the convolution of v and x.
}
\usage{
convmtx(v,n)
}
\arguments{
  \item{v}{ A filter.
}
  \item{n}{ Size of the convolution matrix.
}
}
\value{
The convolution matrix, A, associated to the filter v such that the product of A and an n-element vector, x, is the convolution of v and x.
}
\references{
S. Achard, M. Clausel, I. Gannaz, F. Roueff (2020). New results on approximate Hilbert pairs of wavelet filters with common factor structure. \emph{Applied and Computational Harmonic Analysis}, Vol 49, N.3, pp 1025-1045. 

}
\author{
Achard, Clausel, Gannaz, Roueff (2017)
}

